feat(lite): extract and update "unreachable hosts modal"
